+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project are documented in this file, based on [Keep 
 - Source file configuration in `spoofaxc.cfg` to be relative to the respective source directory, instead of relative to the root directory.
 - Missing source directories or files not giving errors in `spoofaxc.cfg`.
 - Serialization failure when an Eclipse resource issue occurred due to `CoreException` not being serializable. Fixed by turning `CoreException` into `SerializableCoreException` which is serializable.
+- No errors being printed when reading configuration failed in the LWB Gradle plugin.
 
 ### Changed
 - Spoofax Eclipse LWB plugin to show an error dialog when building a language fails due to errors in language definition files.